create separate component for GenerateMessages
