+++ src/pages/conversation/Conversation.css
@@ -175,12 +175,12 @@
   flex-direction: column;
 }
 
-.view-or-generate-btns {
+.view-or-generate-rizz-btns {
   display: flex;
   margin: 0 -12px;
 }
 
-.view-or-generate-btns > * {
+.view-or-generate-rizz-btns > * {
   width: 100%;
   font-size: 18px;
   margin: 0 12px;
